backend/requirements.txt
passlib>=1.7.4
bcrypt==4.0.1
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
alembic==1.13.3
# For Postgres (optional). Remove if not needed.
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib==1.7.4
